+++ docs/examples/gui/01-adaptive_pendulum.py
@@ -1,3 +1,4 @@
+import logging
 import numpy as np
 
 import nengo
@@ -6,7 +7,7 @@ from nengo_fpga.networks import FpgaPesEnsembleNetwork
 
 # Set the nengo logging level to 'info' to display all of the information
 # coming back over the ssh connection.
-nengo.utils.logging.log("info")
+logging.basicConfig(format="%(levelname)s:%(message)s", level=logging.INFO)
 
 # ---------------- BOARD SELECT ----------------------- #
 # Change this to your desired device name
@@ -75,7 +76,9 @@ class Pendulum:
             <line x1="{x1}" y1="{y1}" x2="{x3}" y2="{y3}" style="stroke:blue"/>
             <line x1="{x1}" y1="{y1}" x2="{x2}" y2="{y2}" style="stroke:black"/>
         </svg>
-        """.format(x1=x1, y1=y1, x2=x2, y2=y2, x3=x3, y3=y3)
+        """.format(
+            x1=x1, y1=y1, x2=x2, y2=y2, x3=x3, y3=y3
+        )
 
 
 class PendulumNetwork(nengo.Network):