+++ CHANGES.rst
@@ -26,11 +26,15 @@ Release History
 - Setup Nengo Bones and remote CI.
   (`#41 <https://github.com/nengo/nengo-fpga/pull/41>`__)
 
+**Changed**
+
+- Compatibility changes for Nengo 3.0.0.
+  (`#44 <https://github.com/nengo/nengo-fpga/pull/44>`__)
+
 **Fixed**
 
 - Fixed code to remove all linter errors.
   (`#45 <https://github.com/nengo/nengo-fpga/pull/45>`__)
-
 
 0.2.1 (September 17, 2019)
 --------------------------